backend: user-service - error handling and custom exceptions, some logging added, frontend updated correspondingly

+++ user-service/src/main/java/com/musicclouds/user/controller/UserController.java
@@ -67,7 +67,7 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateUser(@PathVariable Long id, @RequestBody User updatedUser) {
+    public ResponseEntity<?> updateUser(@PathVariable Integer id, @RequestBody User updatedUser) {
         log.info("Starting updateUser({})...", id);
         return userService.updateUser(id, updatedUser)
                 .map(user -> ResponseEntity.ok().body(user))
